Refactor memory management into separate module
